requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
python-dotenv
python-multipart
openai
pymongo
motor 
pandas
scikit-learn
faiss-cpu 
SQLAlchemy